refactor(global-state): rename type file
